+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 3.0.0
 
+- Air Defense focus mode. (Currently anti-air only, so bombers etc. can't be selected)
 - Rewrite atlasMessage using debugger API
 - Note: circa Titans, units are no longer drawn when they would normally be hidden by their icon.